Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/workers/interfaces/DedicatedWorkerGlobalScope/postMessage/second-argument-null-expected.txt


FAIL Using null in postMessage's second argument assert_equals: expected (boolean) true but got (number) 1
